module4-task1 - more details - убрала модули, отбила const пустой строкой
