+++ index.html
@@ -252,6 +252,6 @@
     </div>
   </template>
 
-  <script src="js/main.js" type="module"></script>
+  <script src="js/main.js"></script>
 </body>
 </html>
